% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTFBoost.R
\name{RTFBoost.predict}
\alias{RTFBoost.predict}
\title{RTFBoost.predict}
\usage{
RTFBoost.predict(model, newx, newz = NULL, newy = NULL, grid, t.range)
}
\arguments{
\item{model}{an object returned by \code{RTFBoost}}

\item{newx}{a matrix or data.frame containing observations of the functional predictor in the new data}

\item{newz}{a  matrix or data.frame containing values of the scalar predictors in the new data (required if \code{z.train} and  \code{z.val} were used to fit \code{model},  defaults to \code{NULL}).}

\item{newy}{a  vector containing values of the scalar response in the new data.  (optional,  defaults to \code{NULL})}

\item{grid}{a vector of time points at which \code{x.test} are evaluated.}

\item{t.range}{a vector provided in the form of c(a, b), where a and b are numerical values specifying the boundaries of the domain of the functional predictor.}
}
\value{
A list with with the following components:

\item{f.new}{predicted values with model at the early stopping iteration using x_test (or x_test and z_test) as the predictors}
\item{err.new}{a vector of test errors before and at the early stopping iteration (returned if newy is provided)}
\item{save.f.new}{a matrix of test function estimates at all iterations (returned if save_f = TRUE in control)}
}
\description{
A function to make predictions and calculate test error given an object returned by TFBoost and test data
}
\details{
A function to make predictions and calculate test error given an object returned by TFBoost and test data
}
\examples{
\dontrun{
data(GED)
n <- nrow(GED$price) 
n0 <- floor( 0.2 * n) 
set.seed(123)
idx_test <- sample(n, n0)
idx_train <- sample((1:n)[-idx_test], floor( 0.6 * n ) )
idx_val <- (1:n)[ -c(idx_test, idx_train) ] 
xtrain <- GED$price[idx_train, ]
ytrain <- GED$demand[idx_train ]
xval <- GED$price[idx_val, ]
yval <- GED$demand[idx_val ]
xtest <- GED$price[idx_test, ]
ytest <- GED$demand[idx_test ]
gg <- 1:24  
tt <- c(0,24) 
niter <- 1000
my.control <- TFBoost.control(make_prediction = TRUE, save_tree = TRUE,
                              tree_control = TREE.control(tree_type  = "B", d = 1, num_dir = 200), 
                             shrinkage = 0.05, nknot = 3, loss = "l2")
model_TFBoost <- TFBoost(x_train = xtrain, y_train = ytrain,  x_val = xval,  y_val = yval, 
                         grid = gg, t_range  = tt, niter = niter, 
                         control = my.control)
predictions<- TFBoost.predict(model_TFBoost, newx = xtest)
} 
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
