% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTFBoost.R
\name{RTFBoost.validation}
\alias{RTFBoost.validation}
\title{Robust tree-based functional boosting with initialization parameters chosen on a validation set}
\usage{
RTFBoost.validation(
  x.train,
  z.train = NULL,
  y.train,
  x.val,
  z.val = NULL,
  y.val,
  x.test,
  z.test = NULL,
  y.test,
  grid,
  t.range,
  max.depth.init.set = c(1, 2, 3, 4),
  min.leafsize.init.set = c(10, 20, 30),
  control = RTFBoost.control(),
  save.all = TRUE
)
}
\arguments{
\item{x.train}{a matrix or data.frame containing observations of the functional predictor in the training data.}

\item{z.train}{a  matrix or data.frame containing values of the scalar predictors in the training data (optional, defaults to \code{NULL}).}

\item{y.train}{a  vector containing values of the scalar response in the training data.}

\item{x.val}{a matrix or data.frame containing observations of the functional predictor in the validation data.}

\item{z.val}{a  matrix or data.frame containing values of the scalar predictors in the validation data (optional, required when \code{z.train} is provided, defaults to \code{NULL}).}

\item{y.val}{a  vector containing values of the scalar response in the validation data.}

\item{x.test}{a matrix or data.frame containing observations of the functional predictor in the test data (required when \code{make.prediction} in control is \code{TRUE}).}

\item{z.test}{a  matrix or data.frame containing values of the scalar predictors in the test data (required when \code{make.prediction} in \code{control} is \code{TRUE}, and \code{z.train} and  \code{z.val}  are provided).}

\item{y.test}{a  vector containing values of the scalar response in the test data.  (optional).}

\item{grid}{a vector of time points at which  \code{x.train} and \code{x.val} are evaluated.}

\item{t.range}{a vector provided in the form of c(a, b), where a and b are numerical values specifying the boundaries of the domain of the functional predictor.}

\item{max.depth.init.set}{a vector of possible values of the maximum depth of the initial LADTree that the algorithm choses from (defaults to c(1,2,3,4)).}

\item{min.leafsize.init.set}{a vector of possible values of the minimum observations per node of the initial LADTree that the algorithm choses from (defaults to c(10,20,30)).}

\item{control}{a named list of control parameters as returned by \link{RTFBoost.control}.}

\item{save.all}{a logical flag: if \code{TRUE} the algorithm saves the predictions for median initialization and LADTree initialization with every combination of \code{max.depth.init.set} and \code{min.leaf.size.init.set} (defaults to \code{TRUE}).}
}
\value{
A list with the following components:

\item{model.best}{an object returned by \link{RTFBoost} that is trained with selected initialization parameters}
\item{param}{a vector of selected initialization parameters (return (0,0) if selected initialization is the median of the training responses)}
\item{errs.val}{a vector of robust prediction errors for validation data from all iterations}
\item{errs.test}{a vector of robust prediction errors for test data from all iterations  before and at the early stopping iteration (returned if \code{make.prediction = TRUE} in \code{control})}
\item{pred.list.train}{a list of predicted values for the training data at the early stopping time  (returned if \code{save.all = TRUE}, one element per combination in the order of median initialization, and then \code{expand.grid(sort(min.leafsize.init.set,TRUE), max_depths= sort(max.depth.init.set))})}
\item{pred.list.val}{a list of predicted values for the validation data at the early stopping time  (returned if \code{save.all = TRUE}, one column per combination)}
\item{pred.list.test}{a list of predicted values for the test data at the early stopping time  (returned if \code{save.all = TRUE} and \code{make.prediction = TRUE} in \code{control}, one element per combination)}
}
\description{
A function to fit TFBoost(RR) (see also \code{\link{RTFBoost}}) where the initialization parameters are chosen
based on the performance on the validation set.
}
\details{
This function runs the TFBoost(RR) algorithm (see \code{\link{RTFBoost}}) on different combinations of the
parameters for the initial fit, and chooses the optimal set based on the performance on the validation set.
}
