% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE.control}
\alias{TREE.control}
\title{Tuning and control parameters for the functional multi-index tree}
\usage{
TREE.control(
  tree.type = "A",
  tree.nindex = 1,
  max.depth = 1,
  num.dir = 20,
  min.leafsize = 2,
  nmulti = 3,
  nscreen = 30
)
}
\arguments{
\item{tree.type}{a character ("A" or "B") denoting Type A or Type B tree (defaults to "A").}

\item{tree.nindex}{an integer specifying the number of indices required by a Type A tree (defaults to 1).}

\item{max.depth}{an integer specifying the maximum depth for a Type A or Type B tree (defaults to 1).}

\item{num.dir}{an integer specifying the number of random directions required by the initial Type B tree (defaults to 20).}

\item{min.leafsize}{an integer specifying the minimum number of observations per leaf node for a Type A or Type B tree (defaults to 2).}

\item{nmulti}{an integer specifying  the number of random initialization points to fit a Type A tree (defaults to 3).}

\item{nscreen}{an integer specifying the number of random points from which to select the initial points to fit a Type A tree (defaults to 30).}
}
\value{
A list of all input parameters
}
\description{
Tuning and control parameters for the functional multi-index tree.
}
\details{
Various tuning and control parameters for the functional multi-index tree algorithm implemented in the
function \code{\link{TREE}}
}
\author{
Xiaomeng Ju, \email{xiaomeng.ju@stat.ubc.ca}
}
