% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE.predict}
\alias{TREE.predict}
\title{TREE.predict}
\usage{
TREE.predict(model, newx, newz)
}
\arguments{
\item{model}{an object returned by \code{TREE}}

\item{newx}{a matrix or data.frame containing basis projections of the functional predictor in the new data.}

\item{newz}{a matrix or data.frame containing values of the scalar predictors in the new data (optional, defaults to \code{NULL}, required if \code{z} was used to fit \code{model}).}
}
\value{
\item{pred}{a vector of predicted values using \code{newx} (or \code{newx} and \code{newz}) as the predictors}
}
\description{
A function to make predictions given an object returned by TREE and test data
}
\details{
A function to make predictions given an object returned by TREE and test data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
