% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE}
\alias{TREE}
\title{Functional multi-index tree}
\usage{
TREE(x, y, z, random.seed, control = TREE.control())
}
\arguments{
\item{x}{a matrix or data.frame containing basis projections of the functional predictor in the training data.}

\item{y}{a vector containing values of the scalar response in the training data (optional, defaults to \code{NULL}).}

\item{z}{a matrix or data.frame containing values of the scalar predictors in the training data (optional, defaults to \code{NULL}).}

\item{random.seed}{an integer as the seed to generate random directions for the Type B tree.}

\item{control}{a named list of control parameters, as returned by \code{\link{TREE.control}}}
}
\value{
A list with the following components:
\item{beta.opt}{a matrix of index coefficients estimated by the Type A tree, returned if \code{control$tree.type = "A"}}
\item{betas.selected}{a matrix of index coefficients selected  by the Type B tree, returned if \code{control$tree.type = "B"}}
\item{pred}{a vector of predicted values for the training data}
\item{tree.model}{an rpart object of the fitted tree}
\item{control}{\code{control} from the input arguments}
}
\description{
This function implements a algorithm for functional multi-index tree.
}
\details{
This function implements a functional multi-index tree algorithm developed based on functions available in the \code{rpart} package.
}
\author{
Xiaomeng Ju, \email{xiaomeng.ju@stat.ubc.ca}
}
